feat: standard numerical inputs
