+++ src/components/input/index.tsx
@@ -1,12 +1,13 @@
-import { FC, useEffect, useState } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 import styled from 'styled-components';
+import { neutralColors, brandColors } from '@giveth/ui-design-system';
 
 interface IHasBG {
-	bg?: string;
+	bg?: false;
 }
 
-const InputContainer = styled.div<IHasBG>`
-	background: ${props => (props.bg ? props.bg : '#310BB5')};
+const InputContainer = styled.div`
+	background: #310bb5;
 	border-radius: 34px;
 	padding: 10px 10px 10px 32px;
 	height: 68px;
@@ -16,9 +17,9 @@ const InputContainer = styled.div<IHasBG>`
 	margin: 8px 0;
 `;
 
-const Input = styled.input<IHasBG>`
+const Input = styled.input`
 	border: 0;
-	background: ${props => (props.bg ? props.bg : '#310BB5')};
+	background: #310bb5;
 	color: white;
 	flex: 1;
 	font-size: 18px;
@@ -34,6 +35,11 @@ const Input = styled.input<IHasBG>`
 		-webkit-appearance: none;
 		margin: 0;
 	}
+`;
+
+const Unit = styled.span`
+	padding-right: 10px;
+	color: #cabaff;
 `;
 
 const Button = styled.button<IHasBG>`
@@ -95,15 +101,100 @@ export const WalletAddressInputWithButton: FC<IWalletAddressInputWithButtonProps
 		);
 	};
 
-const Unit = styled.span`
-	padding-right: 10px;
-	color: #cabaff;
+const inputRegex = RegExp(`^\\d*(?:\\\\[.])?\\d*$`); // match escaped "." characters via in a non-capturing group
+
+function escapeRegExp(string: string): string {
+	return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
+}
+
+// Shamelessly copied from Sushiswap front-end
+const BaseInput = React.memo(
+	({
+		value,
+		onUserInput,
+		placeholder,
+		...rest
+	}: {
+		value: string;
+		onUserInput: (input: string) => void;
+		error?: boolean;
+		fontSize?: string;
+		align?: 'right' | 'left';
+	} & Omit<React.HTMLProps<HTMLInputElement>, 'ref' | 'onChange' | 'as'>) => {
+		const enforcer = (nextUserInput: string) => {
+			if (
+				nextUserInput === '' ||
+				inputRegex.test(escapeRegExp(nextUserInput))
+			) {
+				onUserInput(nextUserInput);
+			}
+		};
+
+		return (
+			<Input
+				{...rest}
+				value={value}
+				onChange={event => {
+					// replace commas with periods, because uniswap exclusively uses period as the decimal separator
+					enforcer(event.target.value.replace(/,/g, '.'));
+				}}
+				// universal input options
+				inputMode='decimal'
+				title='Token Amount'
+				autoComplete='off'
+				autoCorrect='off'
+				// text-specific options
+				type='text'
+				pattern='^[0-9]*[.,]?[0-9]*$'
+				placeholder={placeholder || '0.0'}
+				min={0}
+				minLength={1}
+				maxLength={79}
+				spellCheck='false'
+			/>
+		);
+	},
+);
+
+BaseInput.displayName = 'BaseInput';
+
+export const NumericalInput = styled(BaseInput)`
+	width: 100%;
+	height: 54px;
+	padding: 15px 16px;
+	margin-top: 10px;
+	margin-bottom: 8px;
+
+	background: ${brandColors.giv[700]};
+	color: ${neutralColors.gray[100]};
+
+	border: 1px solid ${brandColors.giv[500]};
+	border-radius: 8px;
+
+	font-family: Red Hat Text;
+	font-style: normal;
+	font-weight: normal;
+	font-size: 16px;
+	line-height: 150%;
+
+	&:focus {
+		outline: none;
+	}
+	&[type='number'] {
+		-moz-appearance: textfield;
+	}
+	&::-webkit-outer-spin-button,
+	&::-webkit-inner-spin-button {
+		-webkit-appearance: none;
+		margin: 0;
+	}
+	${props => (props.disabled ? `color: ${brandColors.giv[300]};` : '')}
 `;
 
 interface IInputWithUnitProps {
 	placeholder?: string;
 	unit: string;
-	value: string | number;
+	value: string;
 	onChange?: any;
 	type?: string;
 }
@@ -117,11 +208,11 @@ export const InputWithUnit: FC<IInputWithUnitProps> = ({
 }) => {
 	return (
 		<InputContainer>
-			<Input
+			<BaseInput
 				type={type}
 				placeholder={placeholder}
 				value={value}
-				onChange={onChange}
+				onUserInput={onChange}
 			/>
 			<Unit>{unit}</Unit>
 		</InputContainer>
